Write a program that reads three edges
Write a program that reads three edges a, b and c for a triangle and computes the perimeter if the input is valid. Otherwise, display that the input is invalid. The input is valid if the sum of every pair of two edges is greater than the remaining edge (perimeter = a+ b +c).