which will ask the user to guess a number
Create a console application which will ask the user to guess a number. The developer of the code will choose a secret number between 0 and 20. If the user enters the correct number, the message “Well done, you guessed correctly!” will appear on the screen. If the secret number is lower than the number entered by the user then show the message “Your guess is too low, please try again”, otherwise show the message “Your guess is too high, please try again”.  The program also displays the correct number.